requirements.txt
Django>=2
djangorestframework>=3
inflection
